requirements.txt
numpy
pygame
pandas
pyopengl
pyrr
glfw
